Move `ResourceMetadata` into its own module.
